third_party/WebKit/LayoutTests/platform/mac/virtual/mojo-loading/http/tests/devtools/application-panel/resources-panel-selection-on-reload-expected.txt

CONSOLE MESSAGE: line 124: InspectorTest.IndexedDB_callback1
Tests Application Panel response to a main frame navigation.

Initial state:
Application
  Manifest
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
    Database1 - http://127.0.0.1:8000
  Web SQL
    database-for-test
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
    credentials - http://127.0.0.1:8000
  Application Cache
Frames
  top
    Scripts
      console-test.js
      indexeddb-test.js
      inspector-test.js
      resources-test.js
    resources-panel-selection-on-reload.html
Selection: ["cookies://http://127.0.0.1:8000","category://Cookies"]
Visible view is a cookie view: true
Page reloaded.
After navigation:
Application
  Manifest
  Service Workers
  Clear storage
Storage
  Local Storage
    http://127.0.0.1:8000
  Session Storage
    http://127.0.0.1:8000
  IndexedDB
    Database1 - http://127.0.0.1:8000
  Web SQL
  Cookies
    http://127.0.0.1:8000
Cache
  Cache Storage
    credentials - http://127.0.0.1:8000
  Application Cache
Frames
  top
    Scripts
      console-test.js
      indexeddb-test.js
      inspector-test.js
      resources-test.js
    resources-panel-selection-on-reload.html
Selection: ["cookies://http://127.0.0.1:8000","category://Cookies"]
Visible view is a cookie view: true
